+++ crates/entrypoint/src/cli/command/azure_devops/license_entitlement/user/revoke/mod.rs
@@ -1,0 +1,2 @@
+mod azure_devops_license_entitlement_user_revoke_cli;
+pub use azure_devops_license_entitlement_user_revoke_cli::*;