+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Rename `AzureDevOpsLicenseKind` to `AzureDevOpsLicenseType`, add stronger typing
 - Fix Microsoft graph helper failing to deserialize error bodies
 - Add cache key support to Microsoft graph helper
+- Add `ct az devops license-entitlement user revoke --user-id {}` <-- currently just shows what group to remove them from without doing the removal
 
 # v0.32.0